-Updated the affinity calculation -fixed the memory dump error in the statistical calculation -the affinity calculation was refined by Ivan Sekaj

+++ implemented/globaff.m
@@ -1,5 +1,5 @@
 function[aaff]=globaff(g,eps)
-aaff = uint8(0);
+aaff = 0;
     for i=1:size(g,1)
         aaff = aaff + sum(sum(abs(repmat(g(i,:),size(g,1),1) - g) < eps,2))/size(g,2) - 1;
     end
